# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Book",
        "description": "desplay books whith gender and author",
        "contact": {
            "email": "adrien.charrier@outlook.fr"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/books": {
            "get": {
                "tags": [
                    "book"
                ],
                "summary": "Get List Of Book",
                "description": "Returns all books.",
                "operationId": "GetAllBooks",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "make a search in book's title",
                        "required": false
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "sort results with the book's title or the number of pages [title|pages]",
                        "required": false
                    },
                    {
                        "name": "filterGenre",
                        "in": "query",
                        "description": "filter results with genre id",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "book"
                ],
                "summary": "New Book",
                "description": "Put a new book in the database.",
                "operationId": "NewBook",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "title",
                        "required": true
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "description",
                        "required": true
                    },
                    {
                        "name": "author_id",
                        "in": "query",
                        "description": "author_id",
                        "required": true
                    },
                    {
                        "name": "publication_year",
                        "in": "query",
                        "description": "publication_year",
                        "required": true
                    },
                    {
                        "name": "pages_nb",
                        "in": "query",
                        "description": "pages_nb",
                        "required": true
                    },
                    {
                        "name": "genres",
                        "in": "query",
                        "description": "genres in json",
                        "required": true,
                        "example": "{id: [2, 4]}"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/books/search": {
            "get": {
                "tags": [
                    "book"
                ],
                "summary": "Get List Of Book",
                "description": "Returns all books whith search in the title or the description.",
                "operationId": "SearchBooks",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "make a search in book's title",
                        "required": false
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "make a search in book's description",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/books/{id}": {
            "get": {
                "tags": [
                    "book"
                ],
                "summary": "Get A Book",
                "description": "Get A Book",
                "operationId": "GetABook",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Book id",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "book"
                ],
                "summary": "Delete A Book",
                "description": "Delete A Book",
                "operationId": "DeleteAnBooks",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Book id",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "book"
                ],
                "summary": "Update Book",
                "description": "Update a book in the database.",
                "operationId": "UpdateBook",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "title",
                        "required": true
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "description",
                        "required": true
                    },
                    {
                        "name": "author_id",
                        "in": "query",
                        "description": "author_id",
                        "required": true
                    },
                    {
                        "name": "publication_year",
                        "in": "query",
                        "description": "publication_year",
                        "required": true
                    },
                    {
                        "name": "pages_nb",
                        "in": "query",
                        "description": "pages_nb",
                        "required": true
                    },
                    {
                        "name": "genres",
                        "in": "query",
                        "description": "genres in json",
                        "required": true,
                        "example": "{id: [2, 4]}"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
    "security": [
        []
    ]
}